fix: remove indirections in protocols:tls:TlsStream
