+++ pingora-core/src/protocols/tls/boringssl_openssl/stream.rs
@@ -14,36 +14,47 @@
 
 //! BoringSSL & OpenSSL TLS stream specific implementation
 
+use crate::tls::hash::MessageDigest;
 use log::warn;
 use std::pin::Pin;
 use std::sync::Arc;
-use tokio::io::{AsyncRead, AsyncWrite};
+use std::task::{Context, Poll};
+use std::time::SystemTime;
+use tokio::io;
+use tokio::io::{AsyncRead, AsyncWrite, ReadBuf};
 
-use pingora_error::{Error, ErrorType::*, OrErr, Result};
-
+use crate::listeners::ALPN;
 use crate::protocols::digest::{GetSocketDigest, SocketDigest, TimingDigest};
 use crate::protocols::raw_connect::ProxyDigest;
 use crate::protocols::tls::SslDigest;
-use crate::protocols::{GetProxyDigest, GetTimingDigest};
+use crate::protocols::{GetProxyDigest, GetTimingDigest, Ssl, UniqueID};
 use crate::tls::error::ErrorStack;
 use crate::tls::ext;
-use crate::tls::tokio_ssl;
 use crate::tls::tokio_ssl::SslStream;
 use crate::tls::{ssl, ssl::SslRef, ssl_sys::X509_V_ERR_INVALID_CALL};
+use pingora_error::{Error, ErrorType::*, OrErr, Result};
 
 #[derive(Debug)]
-pub struct InnerStream<T>(pub(crate) tokio_ssl::SslStream<T>);
+pub struct TlsStream<T> {
+    pub(crate) stream: SslStream<T>,
+    pub(super) digest: Option<Arc<SslDigest>>,
+    pub(super) timing: TimingDigest,
+}
 
-impl<T: AsyncRead + AsyncWrite + Unpin> InnerStream<T> {
+impl<T: AsyncRead + AsyncWrite + Unpin> TlsStream<T> {
     /// Create a new TLS connection from the given `stream`
     ///
     /// The caller needs to perform [`Self::connect()`] or [`Self::accept()`] to perform TLS
     /// handshake after.
     pub(crate) fn new(ssl: ssl::Ssl, stream: T) -> Result<Self> {
-        let ssl = SslStream::new(ssl, stream)
+        let stream = SslStream::new(ssl, stream)
             .explain_err(TLSHandshakeFailure, |e| format!("tls.rs stream error: {e}"))?;
 
-        Ok(InnerStream(ssl))
+        Ok(TlsStream {
+            stream,
+            timing: Default::default(),
+            digest: None,
+        })
     }
 
     #[inline]
@@ -55,12 +66,16 @@ impl<T: AsyncRead + AsyncWrite + Unpin> InnerStream<T> {
     }
 }
 
-impl<T: AsyncRead + AsyncWrite + Unpin + Send> InnerStream<T> {
+impl<T: AsyncRead + AsyncWrite + Unpin + Send> TlsStream<T> {
     /// Connect to the remote TLS server as a client
     pub(crate) async fn connect(&mut self) -> Result<()> {
         Self::clear_error();
-        match Pin::new(&mut self.0).connect().await {
-            Ok(_) => Ok(()),
+        match Pin::new(&mut self.stream).connect().await {
+            Ok(_) => {
+                self.timing.established_ts = SystemTime::now();
+                self.digest = self.digest();
+                Ok(())
+            }
             Err(err) => self.transform_ssl_error(err),
         }
     }
@@ -68,18 +83,20 @@ impl<T: AsyncRead + AsyncWrite + Unpin + Send> InnerStream<T> {
     /// Finish the TLS handshake from client as a server
     pub(crate) async fn accept(&mut self) -> Result<()> {
         Self::clear_error();
-        match Pin::new(&mut self.0).accept().await {
-            Ok(_) => Ok(()),
+        match Pin::new(&mut self.stream).accept().await {
+            Ok(_) => {
+                self.timing.established_ts = SystemTime::now();
+                self.digest = self.digest();
+                Ok(())
+            }
             Err(err) => self.transform_ssl_error(err),
         }
     }
 
     pub(crate) fn digest(&mut self) -> Option<Arc<SslDigest>> {
-        Some(Arc::new(SslDigest::from_ssl(self.0.ssl())))
+        Some(Arc::new(SslDigest::from_ssl(self.stream.ssl())))
     }
-}
 
-impl<T: AsyncRead + AsyncWrite + Unpin> InnerStream<T> {
     fn transform_ssl_error(&self, e: ssl::Error) -> Result<()> {
         let context = format!("ssl::ErrorCode: {:?}", e.code());
         if ext::is_suspended_for_cert(&e) {
@@ -102,7 +119,7 @@ impl<T: AsyncRead + AsyncWrite + Unpin> InnerStream<T> {
                         ssl.verify_result().map_err(|e| e.as_raw())
                     }
 
-                    match verify_result(self.0.ssl()) {
+                    match verify_result(self.stream.ssl()) {
                         Ok(()) => Error::e_explain(TLSHandshakeFailure, context),
                         // X509_V_ERR_INVALID_CALL in case verify result was never set
                         Err(X509_V_ERR_INVALID_CALL) => {
@@ -118,32 +135,139 @@ impl<T: AsyncRead + AsyncWrite + Unpin> InnerStream<T> {
     }
 }
 
-impl<S> GetSocketDigest for InnerStream<S>
+impl<T> AsyncRead for TlsStream<T>
+where
+    T: AsyncRead + AsyncWrite + Unpin,
+{
+    fn poll_read(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        buf: &mut ReadBuf<'_>,
+    ) -> Poll<io::Result<()>> {
+        Self::clear_error();
+        Pin::new(&mut self.stream).poll_read(cx, buf)
+    }
+}
+
+impl<T> AsyncWrite for TlsStream<T>
+where
+    T: AsyncRead + AsyncWrite + Unpin,
+{
+    fn poll_write(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context,
+        buf: &[u8],
+    ) -> Poll<io::Result<usize>> {
+        Self::clear_error();
+        Pin::new(&mut self.stream).poll_write(cx, buf)
+    }
+
+    fn poll_flush(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<io::Result<()>> {
+        Self::clear_error();
+        Pin::new(&mut self.stream).poll_flush(cx)
+    }
+
+    fn poll_shutdown(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<io::Result<()>> {
+        Self::clear_error();
+        Pin::new(&mut self.stream).poll_shutdown(cx)
+    }
+
+    fn poll_write_vectored(
+        mut self: Pin<&mut Self>,
+        cx: &mut Context<'_>,
+        bufs: &[std::io::IoSlice<'_>],
+    ) -> Poll<io::Result<usize>> {
+        Self::clear_error();
+        Pin::new(&mut self.stream).poll_write_vectored(cx, bufs)
+    }
+
+    fn is_write_vectored(&self) -> bool {
+        true
+    }
+}
+
+impl<T> Ssl for TlsStream<T> {
+    fn get_ssl(&self) -> Option<&ssl::SslRef> {
+        Some(self.stream.ssl())
+    }
+
+    fn get_ssl_digest(&self) -> Option<Arc<SslDigest>> {
+        self.digest.clone()
+    }
+
+    fn selected_alpn_proto(&self) -> Option<ALPN> {
+        let ssl = self.stream.ssl();
+        ALPN::from_wire_selected(ssl.selected_alpn_protocol()?)
+    }
+}
+
+impl SslDigest {
+    pub fn from_ssl(ssl: &SslRef) -> Self {
+        let cipher = match ssl.current_cipher() {
+            Some(c) => c.name(),
+            None => "",
+        };
+
+        let (cert_digest, org, sn) = match ssl.peer_certificate() {
+            Some(cert) => {
+                let cert_digest = match cert.digest(MessageDigest::sha256()) {
+                    Ok(c) => c.as_ref().to_vec(),
+                    Err(_) => Vec::new(),
+                };
+                (
+                    cert_digest,
+                    crate::utils::tls::boringssl_openssl::get_x509_organization(&cert),
+                    crate::utils::tls::boringssl_openssl::get_x509_serial(&cert).ok(),
+                )
+            }
+            None => (Vec::new(), None, None),
+        };
+
+        SslDigest {
+            cipher,
+            version: ssl.version_str(),
+            organization: org,
+            serial_number: sn,
+            cert_digest,
+        }
+    }
+}
+
+impl<S> GetSocketDigest for TlsStream<S>
 where
     S: GetSocketDigest,
 {
     fn get_socket_digest(&self) -> Option<Arc<SocketDigest>> {
-        self.0.get_ref().get_socket_digest()
+        self.stream.get_ref().get_socket_digest()
     }
     fn set_socket_digest(&mut self, socket_digest: SocketDigest) {
-        self.0.get_mut().set_socket_digest(socket_digest)
+        self.stream.get_mut().set_socket_digest(socket_digest)
     }
 }
 
-impl<S> GetTimingDigest for InnerStream<S>
+impl<S> GetTimingDigest for TlsStream<S>
 where
     S: GetTimingDigest,
 {
     fn get_timing_digest(&self) -> Vec<Option<TimingDigest>> {
-        self.0.get_ref().get_timing_digest()
+        self.stream.get_ref().get_timing_digest()
     }
 }
 
-impl<S> GetProxyDigest for InnerStream<S>
+impl<S> GetProxyDigest for TlsStream<S>
 where
     S: GetProxyDigest,
 {
     fn get_proxy_digest(&self) -> Option<Arc<ProxyDigest>> {
-        self.0.get_ref().get_proxy_digest()
+        self.stream.get_ref().get_proxy_digest()
+    }
+}
+
+impl<T> UniqueID for TlsStream<T>
+where
+    T: UniqueID,
+{
+    fn id(&self) -> i32 {
+        self.stream.get_ref().id()
     }
 }
